refactoring and remove session stuff
