--- requirements.txt
@@ -1,13 +1,0 @@
-certifi==2020.11.8
-chardet==3.0.4
-click==7.1.2
-Flask==1.1.2
-idna==2.10
-itsdangerous==1.1.0
-Jinja2==2.11.2
-MarkupSafe==1.1.1
-python-dotenv @ file:///Users/Jordan/Library/Caches/pypoetry/artifacts/10/87/64/df36e1a8f4db1b5db8eadd5084df803a683599a450a09e22fb1a7b95db/python_dotenv-0.14.0-py2.py3-none-any.whl
-requests==2.25.0
-todo-app==0.1.0
-urllib3==1.26.1
-Werkzeug==1.0.1
