+++ todo_app/app.py
@@ -1,8 +1,6 @@
 from flask import Flask, render_template, request, redirect
-from todo_app.data.session_items import get_items, add_item, save_item, get_item, delete_item
 from todo_app.flask_config import Config
-from todo_app.service.trello_service import create_to_do, delete_card, move_card_to_complete, \
-    move_card_to_in_progress, move_card_to_to_do, get_cards, fetch_updated_cards, sort_cards, get_sorted_status
+import todo_app.service.trello_service as trello
 
 app = Flask(__name__)
 app.config.from_object(Config)
@@ -13,78 +11,50 @@ def heartbeat():
     return 'Flask Server is Running'
 
 
-@app.route('/session')
-def session_cards():
-    return render_template('index.html', items=get_items())
-
-
 @app.route('/')
 def trello_cards():
-    fetch_updated_cards()
-    return render_template('trello_index.html', cards=get_cards(), sort_status = get_sorted_status())
-
-
-@app.route('/session_add', methods=['POST'])
-def session_add_to_do():
-    add_item(request.form['title'])
-    return redirect('/session')
-
-
-@app.route('/session_complete', methods=['POST'])
-def session_update_to_do():
-    id = request.form['id']
-    item = get_item(id)
-    item['status'] = "Complete"
-    save_item(item)
-    return redirect('/session')
-
-
-@app.route('/session_delete', methods=['POST'])
-def session_delete_to_do():
-    id = request.form['id']
-    item = get_item(id)
-    delete_item(item)
-    return redirect('/session')
+    trello.fetch_updated_cards()
+    return render_template('index.html', cards=trello.get_cards(), sort_status = trello.get_sorted_status())
 
 
 @app.route('/add', methods=['POST'])
 def add_to_do():
-    create_to_do(request.form['title'])
+    trello.create_to_do(request.form['title'])
     return redirect('/')
 
 
 @app.route('/complete', methods=['POST'])
 def complete_to_do():
     card_id = request.form['id']
-    move_card_to_complete(card_id)
+    trello.move_card_to_complete(card_id)
     return redirect('/')
 
 
 @app.route('/start', methods=['POST'])
 def start_to_do():
     card_id = request.form['id']
-    move_card_to_in_progress(card_id)
+    trello.move_card_to_in_progress(card_id)
     return redirect('/')
 
 
 @app.route('/reset', methods=['POST'])
 def reset_to_do():
     card_id = request.form['id']
-    move_card_to_to_do(card_id)
+    trello.move_card_to_to_do(card_id)
     return redirect('/')
 
 
 @app.route('/delete', methods=['POST'])
 def delete_to_do():
     card_id = request.form['id']
-    delete_card(card_id)
+    trello.delete_card(card_id)
     return redirect('/')
 
 
 @app.route('/sort', methods=['POST'])
 def sort():
     ascending = request.form['id']
-    sort_cards(ascending)
+    trello.sort_cards(ascending)
     return redirect('/')
