+++ todo_app/data/Card.py
@@ -1,0 +1,6 @@
+class Card:
+
+    def __init__(self, card_id, description, status):
+        self.card_id = card_id
+        self.description = description
+        self.status = status
