+++ todo_app/templates/index.html
@@ -2,63 +2,102 @@
 {% block title %}To-Do App{% endblock %}
 
 {% block content %}
-  <div class="jumbotron">
-    <h1 class="display-4">To-Do App</h1>
-    <p class="lead">Just another to-do app.</p>
-  </div>
+    <div class="jumbotron">
+        <h1 class="display-4">To-Do App</h1>
+        <p class="lead">Just another to-do app. Trello Version</p>
+    </div>
 
-  <div class="row">
-    <div class="col-md-12">
-        <h2>Add a New To-Do</h2>
-        <form class="form-inline" action="/session_add" method="post">
-            <div class="form-group ml-2">
-                <input required type="text" class="form-control" name="title" placeholder="My New To-Do">
-            </div>
-            <button type="submit" class="btn btn-primary ml-2">Add</button>
-        </form>
-        <h2>Items</h2>
-
-        <ul class="list-group mb-4">
-            <table class="table">
-                <thead>
-                <tr>
-                    <th scope="col">id</th>
-                    <th scope="col">Title</th>
-                    <th scope="col">Status</th>
-                    <th scope="col"></th>
-                </tr>
-                </thead>
-                <tbody>
-                {% for item in items %}
+    <div class="row">
+        <div class="col-md-12">
+            <h2>Add a New To-Do</h2>
+            <form class="form-inline" action="/add" method="post">
+                <div class="form-group ml-2">
+                    <input required type="text" class="form-control" name="title" placeholder="My New To-Do">
+                </div>
+                <button type="submit" class="btn btn-primary ml-2">Add</button>
+            </form>
+            <h2>Items</h2>
+            <ul class="list-group mb-4">
+                <table class="table">
+                    <thead>
                     <tr>
-                        <td>{{ item.id }}</td>
-                        <td
-                                {% if  item.status != 'Not Started'  %}
-                                style="text-decoration: line-through;"
-                                {% endif %}
-                        >{{ item.title }}</td>
-                        {% if  item.status == 'Not Started'  %}
+                        <th scope="col"></th>
+                        <th scope="col">Description</th>
+                        <th scope="col">
+                            {% if sort_status=="ASCENDING" %}
+                                <form action="/sort" method="post">
+                                    <input hidden name="id" value="DESCENDING">
+                                    Status
+                                    <button type="submit" class="btn btn-link btn-sm" style="width:50px">↑</button>
+                                </form>
+                            {% else %}
+                                <form action="/sort" method="post">
+                                    <input hidden name="id" value="ASCENDING">
+                                    Status
+                                    <button type="submit" class="btn btn-link btn-sm" style="width:50px">↓</button>
+                                </form>
+                            {% endif %}
+                        </th>
+                        <th scope="col"></th>
+                        <th scope="col"></th>
+                        <th scope="col"></th>
+                    </tr>
+                    </thead>
+
+                    <tbody>
+                    {% for card in cards %}
+                        <tr>
                             <td>
-                                <form action="/session_complete" method="post">
-                                    <input hidden name="id" value={{ item.id }}>
-                                    <button type="submit" class="btn btn-primary btn-sm">Mark Complete</button>
+                                <form action="/delete" method="post">
+                                    <input hidden name="id" value={{ card.card_id }}>
+                                    <button type="submit" class="btn btn-danger btn-sm">X</button>
                                 </form>
                             </td>
-                        {% else %}
-                            <td style="color: green;font-weight: bold;">{{ item.status }}</td>
-                        {% endif %}
-                    <td>
-                        <form action="/session_delete" method="post">
-                            <input hidden name="id" value={{ item.id }}>
-                            <button type="submit" class="btn btn-danger btn-sm">Remove</button>
-                        </form>
-                    </td>
-                    </tr>
-
-                {% endfor %}
-                </tbody>
-            </table>
-      </ul>
+                            <td
+                                {% if  card.status == 'Done'  %}
+                                    style="text-decoration: line-through;"
+                                {% endif %}
+                            >{{ card.description }}</td>
+                            <td
+                                    {% if  card.status == 'Done'  %}
+                                        style="color: #28a745;font-weight: bold;"
+                                    {% endif %}
+                                    {% if  card.status == 'To-Do'  %}
+                                        style="color: #dc3545;font-weight: bold;"
+                                    {% endif %}
+                                    {% if  card.status == 'In Progress'  %}
+                                        style="color: #007bff;font-weight: bold;"
+                                    {% endif %}
+                            >{{ card.status }}</td>
+                            {% if card.status!="To-Do" %}
+                                <td>
+                                    <form action="/reset" method="post">
+                                        <input hidden name="id" value={{ card.card_id }}>
+                                        <button type="submit" class="btn btn-secondary btn-sm">Move to To-Do</button>
+                                    </form>
+                                </td>
+                            {% endif %}
+                            {% if card.status!="In Progress" %}
+                                <td>
+                                    <form action="/start" method="post">
+                                        <input hidden name="id" value={{ card.card_id }}>
+                                        <button type="submit" class="btn btn-primary btn-sm">Move to In Progress</button>
+                                    </form>
+                                </td>
+                            {% endif %}
+                            {% if card.status!="Done" %}
+                                <td>
+                                    <form action="/complete" method="post">
+                                        <input hidden name="id" value={{ card.card_id }}>
+                                        <button type="submit" class="btn btn-success btn-sm">Move to Complete</button>
+                                    </form>
+                                </td>
+                            {% endif %}
+                        </tr>
+                    {% endfor %}
+                    </tbody>
+                </table>
+            </ul>
+        </div>
     </div>
-  </div>
 {% endblock %}
